requirements.txt
load_dotenv
python-binance
pandas
numpy
ta
matplotlib
scikit-learn
xgboost
websocket-client
joblib
